+++ src/test/suite/fixtures/missingName.md
@@ -1,11 +1,7 @@
 ---
-<<<<<<< HEAD
 type: openRespons
 notes:
 rows: d
-=======
-type: noResponse
->>>>>>> origin/separating-tests
 ---
 
 Please describe the chair you are sitting in.
